+++ tests/testthat/test_PipeOpFDASmooth.R
@@ -1,4 +1,4 @@
-test_that("PipeOpSmooth - basic properties", {
+test_that("PipeOpFDASmooth - basic properties", {
   pop = po("fda.smooth")
   expect_pipeop(pop)
   expect_identical(pop$id, "fda.smooth")